requirements.txt
# Python version requirement: >=3.11
# Core dependencies
python-dotenv>=1.0.0

# LangChain ecosystem - compatible versions
langchain-core>=0.3.74,<1.0.0
langchain>=0.3.0,<1.0.0

# LangGraph for agent workflows
langgraph>=0.2.0
langgraph-checkpoint>=2.1.1

# Anthropic API integration - latest compatible versions
langchain-anthropic>=0.3.19,<1.0.0
anthropic>=0.64.0,<1.0.0

# Modern Pydantic v2
pydantic>=2.7.4,<3.0.0

# HTTP and async dependencies
httpx>=0.27.0
uvicorn>=0.30.0
sse-starlette>=2.0.0

#mcp dependendencies
mcp>=0.3.0,<1.0.0
fastmcp
aiocache
aiolimiter

# Remove built-in modules: json, asyncio (they're part of Python standard library)

# Optional: Additional LangChain integrations (uncomment if needed)
# langchain-openai>=0.3.0,<1.0.0
# langchain-community>=0.3.0,<1.0.0

# Development and testing (optional)
# pytest>=8.0.0
# black>=24.0.0
# ruff>=0.6.0
# mypy>=1.11.0

# LangGraph Studio (optional - install separately)
# langgraph-studio